VideoCore: Split shader regs from Regs struct
